+++ code/lib/plotting/__init__.py
@@ -1,3 +1,4 @@
 from .single_model_single_variation import *
 from .single_model_multi_variations import *
 from .multi_model_plots import *
+from .utils import *
